Time series clean up

+++ zebedee-cms/src/main/java/com/github/onsdigital/zebedee/model/PathUtils.java
@@ -152,7 +152,7 @@ public class PathUtils {
 
         try (DirectoryStream<Path> stream = Files.newDirectoryStream(sourceDirectory)) {
             for (Path entry : stream) {
-                if (!Files.isDirectory(entry)) {
+                if (!Files.isDirectory(entry) && Files.exists(entry)) {
                     Files.delete(entry);
                 }
             }
